requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
anthropic>=0.39.0
hyperspell>=0.28.0